+++ src/quickPicks/quickPickSetupVscodeOpenai.ts
@@ -105,22 +105,17 @@ export async function quickPickSetupVscodeOpenai(
 
   await collectInputs()
   const accessToken = await getApiKey()
-  if (accessToken) {
-    const config: IConfigurationService = {
-      serviceProvider: 'VSCode-OpenAI',
-      baseUrl: `https://api.arbs.io/openai/inference/v1`,
-      defaultModel: 'gpt-35-turbo',
-      embeddingModel: 'text-embedding-ada-002',
-      azureDeployment: 'gpt-35-turbo',
-      embeddingsDeployment: 'text-embedding-ada-002',
-      azureApiVersion: '2023-05-15',
-    }
-    await SecretStorageService.instance.setAuthApiKey(accessToken)
-    await ConfigurationService.loadConfigurationService(config)
-    await verifyApiKey()
-    ExtensionStatusBarItem.instance.showStatusBarInformation(
-      'vscode-openai',
-      ''
-    )
+  if (!accessToken) return
+  const config: IConfigurationService = {
+    serviceProvider: 'VSCode-OpenAI',
+    baseUrl: `https://api.arbs.io/openai/inference/v1`,
+    defaultModel: 'gpt-35-turbo',
+    embeddingModel: 'text-embedding-ada-002',
+    azureDeployment: 'gpt-35-turbo',
+    embeddingsDeployment: 'text-embedding-ada-002',
+    azureApiVersion: '2023-05-15',
   }
+  await SecretStorageService.instance.setAuthApiKey(accessToken)
+  await ConfigurationService.loadConfigurationService(config)
+  ExtensionStatusBarItem.instance.showStatusBarInformation('vscode-openai', '')
 }